+++ src/PSCompression/Commands/CompressZipArchiveCommand.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.IO.Compression;
 using System.Management.Automation;
 using PSCompression.Extensions;
+using static PSCompression.Exceptions.ExceptionHelpers;
 
 namespace PSCompression;
 
@@ -98,19 +99,15 @@ public sealed class CompressZipArchiveCommand : PSCmdlet, IDisposable
         }
         catch (Exception e)
         {
-            ThrowTerminatingError(ExceptionHelpers.StreamOpenError(Destination, e));
+            ThrowTerminatingError(StreamOpenError(Destination, e));
         }
     }
 
     protected override void ProcessRecord()
     {
-        if (_zip is null)
-        {
-            return;
-        }
+        Dbg.Assert(_zip is not null);
 
         _queue.Clear();
-
         foreach (string path in _paths.NormalizePath(_isLiteral, this))
         {
             if (!path.IsArchive())
@@ -120,7 +117,7 @@ public sealed class CompressZipArchiveCommand : PSCmdlet, IDisposable
             }
 
             FileInfo file = new(path);
-            if (Update.IsPresent && TryGetEntry(_zip, file.Name, out ZipArchiveEntry? entry))
+            if (Update.IsPresent && _zip.TryGetEntry(file.Name, out ZipArchiveEntry? entry))
             {
                 UpdateEntry(file, entry);
                 continue;
@@ -142,7 +139,7 @@ public sealed class CompressZipArchiveCommand : PSCmdlet, IDisposable
 
             string relative = current.RelativeTo(length);
 
-            if (!Update.IsPresent || !TryGetEntry(zip, relative, out _))
+            if (!Update.IsPresent || !zip.TryGetEntry(relative, out _))
             {
                 zip.CreateEntry(current.RelativeTo(length));
             }
@@ -157,7 +154,7 @@ public sealed class CompressZipArchiveCommand : PSCmdlet, IDisposable
             }
             catch (Exception e)
             {
-                WriteError(ExceptionHelpers.EnumerationError(current, e));
+                WriteError(EnumerationError(current, e));
                 continue;
             }
 
@@ -178,7 +175,7 @@ public sealed class CompressZipArchiveCommand : PSCmdlet, IDisposable
 
                 relative = file.RelativeTo(length);
 
-                if (Update.IsPresent && TryGetEntry(zip, relative, out ZipArchiveEntry? entry))
+                if (Update.IsPresent && zip.TryGetEntry(relative, out ZipArchiveEntry? entry))
                 {
                     UpdateEntry(file, entry);
                     continue;
@@ -188,9 +185,6 @@ public sealed class CompressZipArchiveCommand : PSCmdlet, IDisposable
             }
         }
     }
-
-    private bool TryGetEntry(ZipArchive zip, string path, out ZipArchiveEntry entry) =>
-        (entry = zip.GetEntry(path)) is not null;
 
     private void CreateEntry(
         FileInfo file,
@@ -212,7 +206,7 @@ public sealed class CompressZipArchiveCommand : PSCmdlet, IDisposable
         }
         catch (Exception e)
         {
-            WriteError(ExceptionHelpers.StreamOpenError(file.FullName, e));
+            WriteError(StreamOpenError(file.FullName, e));
         }
     }
 
@@ -239,7 +233,7 @@ public sealed class CompressZipArchiveCommand : PSCmdlet, IDisposable
         }
         catch (Exception e)
         {
-            WriteError(ExceptionHelpers.StreamOpenError(file.FullName, e));
+            WriteError(StreamOpenError(file.FullName, e));
         }
     }