+++ src/PSCompression/Commands/RemoveZipEntryCommand.cs
@@ -1,6 +1,7 @@
 using System;
 using System.IO.Compression;
 using System.Management.Automation;
+using static PSCompression.Exceptions.ExceptionHelpers;
 
 namespace PSCompression;
 
@@ -29,7 +30,7 @@ public sealed class RemoveZipEntryCommand : PSCmdlet, IDisposable
             }
             catch (Exception e)
             {
-                WriteError(ExceptionHelpers.ZipOpenError(entry.Source, e));
+                WriteError(ZipOpenError(entry.Source, e));
             }
         }
     }