+++ src/PSCompression/Commands/SetZipEntryContentCommand.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Management.Automation;
 using System.Text;
+using static PSCompression.Exceptions.ExceptionHelpers;
 
 namespace PSCompression;
 
@@ -58,8 +59,7 @@ public sealed class SetZipEntryContentCommand : PSCmdlet, IDisposable
         }
         catch (Exception e)
         {
-            ThrowTerminatingError(
-                ExceptionHelpers.StreamOpenError(SourceEntry, e));
+            ThrowTerminatingError(StreamOpenError(SourceEntry, e));
         }
     }
 
@@ -67,10 +67,7 @@ public sealed class SetZipEntryContentCommand : PSCmdlet, IDisposable
     {
         try
         {
-            if (_zipWriter is null)
-            {
-                return;
-            }
+            Dbg.Assert(_zipWriter is not null);
 
             if (AsByteStream.IsPresent)
             {
@@ -87,14 +84,15 @@ public sealed class SetZipEntryContentCommand : PSCmdlet, IDisposable
         }
         catch (Exception e)
         {
-            ThrowTerminatingError(
-                ExceptionHelpers.WriteError(SourceEntry, e));
+            ThrowTerminatingError(ZipWriteError(SourceEntry, e));
         }
     }
 
     protected override void EndProcessing()
     {
-        if (!PassThru.IsPresent || _zipWriter is null)
+        Dbg.Assert(_zipWriter is not null);
+
+        if (!PassThru.IsPresent)
         {
             return;
         }
@@ -107,8 +105,7 @@ public sealed class SetZipEntryContentCommand : PSCmdlet, IDisposable
         }
         catch (Exception e)
         {
-            ThrowTerminatingError(
-                ExceptionHelpers.StreamOpenError(SourceEntry, e));
+            ThrowTerminatingError(StreamOpenError(SourceEntry, e));
         }
     }